LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/rendering/widgets/button-layout/flex-expected.txt


1
2
3
4
 
1
2
3
4

1
2
3
4
 
abc

abc

PASS inline-flex 
PASS flex 
FAIL align-items:flex-start should work assert_equals: expected 116 but got 101
FAIL align-items:stretch should work assert_equals: expected 18 but got 48
